+++ src/Portenta_H7_AsyncHTTPRequest.hpp
@@ -15,9 +15,10 @@
   as published bythe Free Software Foundation, either version 3 of the License, or (at your option) any later version.
   This program is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY; without even the implied warranty of
   MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU General Public License for more details.
-  You should have received a copy of the GNU General Public License along with this program.  If not, see <https://www.gnu.org/licenses/>.  
+  You should have received a copy of the GNU General Public License along with this program.
+  If not, see <https://www.gnu.org/licenses/>.  
  
-  Version: 1.3.0
+  Version: 1.3.1
   
   Version Modified By   Date      Comments
   ------- -----------  ---------- -----------
@@ -25,6 +26,7 @@
   1.1.0    K Hoang     30/12/2021 Fix `multiple-definitions` linker error
   1.2.0    K Hoang     24/01/2022 Enable compatibility with old code to include only AsyncHTTPRequest_Generic.h
   1.3.0    K Hoang     02/09/2022 Fix bug. Improve debug messages. Optimize code
+  1.3.1    K Hoang     18/10/2022 Not try to reconnect to the same host:port after connected
  *****************************************************************************************************************************/
 
 #pragma once
@@ -59,13 +61,13 @@
   #define SHIELD_TYPE           "Portenta_H7 WiFi"
 #endif
 
-#define PORTENTA_H7_ASYNC_HTTP_REQUEST_VERSION            "Portenta_H7_AsyncHTTPRequest v1.3.0"
+#define PORTENTA_H7_ASYNC_HTTP_REQUEST_VERSION            "Portenta_H7_AsyncHTTPRequest v1.3.1"
 
 #define PORTENTA_H7_ASYNC_HTTP_REQUEST_VERSION_MAJOR      1
 #define PORTENTA_H7_ASYNC_HTTP_REQUEST_VERSION_MINOR      3
-#define PORTENTA_H7_ASYNC_HTTP_REQUEST_VERSION_PATCH      0
+#define PORTENTA_H7_ASYNC_HTTP_REQUEST_VERSION_PATCH      1
 
-#define PORTENTA_H7_ASYNC_HTTP_REQUEST_VERSION_INT        1003000
+#define PORTENTA_H7_ASYNC_HTTP_REQUEST_VERSION_INT        1003001
 
 #include "Portenta_H7_AsyncTCP.h"